ocr/requirements.txt
minio==7.2.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0
Pillow==10.1.0
numpy==1.24.3

